# 5.6.7 - 修复预签名徒步没有透传的bug -  修复swift工程中使用sdk的若干警告 - 修复文件过大，int越界的bug - 支持Read_ACP和Write_ACP枚举 \ - 其他bug修复

+++ QCloudCOSXML/Classes/Manager/model/CASJobParameters.m
@@ -64,7 +64,7 @@ NS_ASSUME_NONNULL_BEGIN
 
         NSString* CASTierenumValue = transfromDic[@"Tier"];
         if (CASTierenumValue && [CASTierenumValue isKindOfClass:[NSString class]] && CASTierenumValue.length > 0) {
-            int value = QCloudCASTierDumpFromString(CASTierenumValue);
+            NSInteger value = QCloudCASTierDumpFromString(CASTierenumValue);
             transfromDic[@"Tier"] = @(value);
         }
     return transfromDic;
